Reapply "add spawn logic / add minotaur and python"
This reverts commit 327b1fe85f5656f0bd2628797ae93dd69f28ebb0.

+++ include/server/game/weaponcollider.hpp
@@ -10,7 +10,7 @@ class WeaponCollider : public Object {
 public:
 
     struct WeaponOptions {
-        WeaponOptions(int damage, float timeUntilAttack, int attackDuration, bool followPlayer) :
+        WeaponOptions(int damage, int timeUntilAttack, int attackDuration, bool followPlayer) :
             damage(damage), timeUntilAttack(timeUntilAttack), attackDuration(attackDuration), followPlayer(followPlayer)
         {}
